refactor(ir): dedup BufferRootCollector into shared util
Delete the dead BufferRootCollector instance in orchestration codegen
(GenerateOrchestration built and visited it but never read its result),
then unify the two remaining implementations into one shared analysis.

- Add include/pypto/ir/transforms/utils/buffer_root_collector.{h,cpp}
  (namespace ir::buffer_root), based on the improved copy that lived in
  fuse_create_assemble_to_slice_pass: SelectReturnRoot matches a call's
  single return value to the Out/InOut arg with the same shape+dtype, so a
  differently-typed InOut scratch is never mistaken for the output buffer.
- FuseCreateAssembleToSlice drops its local copy and uses the shared class.
- DeriveCallDirections uses the shared class instead of the codegen copy;
  single-return root selection now goes through SelectReturnRoot.
- Remove the naive BufferRootCollector from codegen orchestration_analysis,
  which also drops the ir -> codegen dependency for this analysis.
