+++ src/codegen/orchestration/orchestration_analysis.cpp
@@ -201,116 +201,6 @@ void OrchestrationInfoCollector::VisitStmt_(const AssignStmtPtr& assign) {
 }
 
 // ---------------------------------------------------------------------------
-// BufferRootCollector
-// ---------------------------------------------------------------------------
-
-BufferRootCollector::BufferRootCollector(ProgramPtr program) : program_(std::move(program)) {}
-
-void BufferRootCollector::Initialize(const std::vector<VarPtr>& params) {
-  for (const auto& param : params) {
-    buffer_roots[param.get()] = param.get();
-  }
-}
-
-void BufferRootCollector::VisitStmt_(const ForStmtPtr& for_stmt) {
-  for (size_t i = 0; i < for_stmt->iter_args_.size(); ++i) {
-    const auto& iter_arg = for_stmt->iter_args_[i];
-    const Var* root = ResolveExpr(iter_arg->initValue_);
-    if (root) {
-      buffer_roots[iter_arg.get()] = root;
-      if (i < for_stmt->return_vars_.size()) {
-        buffer_roots[for_stmt->return_vars_[i].get()] = root;
-      }
-    }
-  }
-  IRVisitor::VisitStmt_(for_stmt);
-}
-
-void BufferRootCollector::VisitStmt_(const WhileStmtPtr& while_stmt) {
-  for (size_t i = 0; i < while_stmt->iter_args_.size(); ++i) {
-    const auto& iter_arg = while_stmt->iter_args_[i];
-    const Var* root = ResolveExpr(iter_arg->initValue_);
-    if (root) {
-      buffer_roots[iter_arg.get()] = root;
-      if (i < while_stmt->return_vars_.size()) {
-        buffer_roots[while_stmt->return_vars_[i].get()] = root;
-      }
-    }
-  }
-  IRVisitor::VisitStmt_(while_stmt);
-}
-
-void BufferRootCollector::VisitStmt_(const AssignStmtPtr& assign) {
-  // Submit funnels through the Call-shaped view so a kernel launch's Out/InOut
-  // roots are tracked identically to a plain Call (results keyed on the stable
-  // binding Var). Builtin tensor.* ops are never Submits, so the early
-  // op-name branches below are unaffected.
-  if (auto call = AsCallOrSubmitView(assign->value_)) {
-    const std::string& op_name = call->op_->name_;
-    if (IsOp(call, "tensor.create") || IsOp(call, "tensor.slice")) {
-      buffer_roots[assign->var_.get()] = assign->var_.get();
-    } else if (IsOp(call, "tensor.assemble")) {
-      if (call->args_.size() == 3) {
-        if (const Var* target_root = ResolveExpr(call->args_[0])) {
-          buffer_roots[assign->var_.get()] = target_root;
-        }
-      }
-    } else if (!IsBuiltinOp(op_name)) {
-      auto out_roots = CollectCallOutputRoots(call);
-      if (As<TupleType>(call->GetType())) {
-        tuple_output_roots_[assign->var_.get()] = std::move(out_roots);
-      } else if (!out_roots.empty() && out_roots[0]) {
-        buffer_roots[assign->var_.get()] = out_roots[0];
-      }
-    }
-  } else if (auto tuple_get = As<TupleGetItemExpr>(assign->value_)) {
-    if (auto tuple_var = AsVarLike(tuple_get->tuple_)) {
-      auto it = tuple_output_roots_.find(tuple_var.get());
-      if (it != tuple_output_roots_.end() && tuple_get->index_ < static_cast<int>(it->second.size()) &&
-          it->second[tuple_get->index_]) {
-        buffer_roots[assign->var_.get()] = it->second[tuple_get->index_];
-      }
-    }
-  } else if (auto src_var = AsVarLike(assign->value_)) {
-    if (const Var* root = ResolveVar(src_var.get())) {
-      buffer_roots[assign->var_.get()] = root;
-    }
-  }
-  IRVisitor::VisitStmt_(assign);
-}
-
-const Var* BufferRootCollector::ResolveVar(const Var* var) const {
-  auto it = buffer_roots.find(var);
-  return it != buffer_roots.end() ? it->second : nullptr;
-}
-
-const Var* BufferRootCollector::ResolveExpr(const ExprPtr& expr) const {
-  if (auto var = AsVarLike(expr)) {
-    return ResolveVar(var.get());
-  }
-  return nullptr;
-}
-
-std::vector<const Var*> BufferRootCollector::CollectCallOutputRoots(const CallPtr& call) const {
-  auto callee = program_->GetFunction(call->op_->name_);
-  if (!callee) return {};
-
-  std::vector<const Var*> roots;
-  for (size_t i = 0; i < callee->param_directions_.size() && i < call->args_.size(); ++i) {
-    if (callee->param_directions_[i] != ParamDirection::Out &&
-        callee->param_directions_[i] != ParamDirection::InOut) {
-      continue;
-    }
-    if (auto arg_var = AsVarLike(call->args_[i])) {
-      roots.push_back(ResolveVar(arg_var.get()));
-    } else {
-      roots.push_back(nullptr);
-    }
-  }
-  return roots;
-}
-
-// ---------------------------------------------------------------------------
 // VarLineageCollector
 // ---------------------------------------------------------------------------
